capitalise table titles and column titles: https://github.com/metanorma/metanorma-itu/issues/496
