+++ lib/isodoc/itu/presentation_xml_convert.rb
@@ -4,6 +4,18 @@ require "isodoc"
 require_relative "../../relaton/render/general"
 require_relative "presentation_bibdata"
 require_relative "presentation_preface"
+require_relative "presentation_ref"
+
+module Nokogiri
+  module XML
+    class Node
+      def traverse_topdown(&block)
+        yield(self)
+        children.each { |j| j.traverse_topdown(&block) }
+      end
+    end
+  end
+end
 
 module IsoDoc
   module ITU
@@ -39,6 +51,24 @@ module IsoDoc
         super
       end
 
+      def table1(elem)
+        elem.xpath(ns("./name | ./thead/tr/th")).each do |n|
+          capitalise_unless_text_transform(n)
+        end
+        super
+      end
+
+      def capitalise_unless_text_transform(elem)
+        css = nil
+        elem.traverse_topdown do |n|
+          n.name == "span" && /text-transform:/.match?(n["style"]) and
+            css = n
+          n.text? && /\S/.match?(n.text) or next
+          css && n.ancestors.include?(css) or n.replace(n.text.capitalize)
+          break
+        end
+      end
+
       def get_eref_linkend(node)
         non_locality_elems(node).select do |c|
           !c.text? || /\S/.match(c)
@@ -50,57 +80,6 @@ module IsoDoc
                                 link, node)
         non_locality_elems(node).each(&:remove)
         node.add_child(link)
-      end
-
-      def bibrenderer
-        ::Relaton::Render::ITU::General.new(language: @lang)
-      end
-
-      def bibrender_formattedref(formattedref, _xml)
-        formattedref << "." unless /\.$/.match?(formattedref.text)
-        id = reference_format_start(formattedref.parent) and
-          formattedref.children.first.previous = id
-      end
-
-      def bibrender_relaton(xml, renderings)
-        f = renderings[xml["id"]][:formattedref]
-        ids = reference_format_start(xml)
-        f &&= "<formattedref>#{ids}#{f}</formattedref>"
-        # retain date in order to generate reference tag
-        keep = "./docidentifier | ./uri | ./note | ./date | ./biblio-tag"
-        xml.children = "#{f}#{xml.xpath(ns(keep)).to_xml}"
-      end
-
-      def multi_bibitem_ref_code(bib)
-        skip = IsoDoc::Function::References::SKIP_DOCID
-        skip1 = "@type = 'metanorma' or @type = 'metanorma-ordinal'"
-        prim = "[@primary = 'true']"
-        id = bib.xpath(ns("./docidentifier#{prim}[not(#{skip} or #{skip1})]"))
-        id.empty? and id = bib.xpath(ns("./docidentifier#{prim}[not(#{skip1})]"))
-        id.empty? and id = bib.xpath(ns("./docidentifier[not(#{skip} or #{skip1})]"))
-        id.empty? and id = bib.xpath(ns("./docidentifier[not(#{skip1})]"))
-        id.empty? and return id
-        id.sort_by { |i| i["type"] == "ITU" ? 0 : 1 }
-      end
-
-      def render_multi_identifiers(ids)
-        ids.map do |id|
-          if id["type"] == "ITU" then doctype_title(id)
-          else
-            docid_prefix(id["type"], id.text.sub(/^\[/, "").sub(/\]$/, ""))
-          end
-        end.join("&#xA0;| ")
-      end
-
-      def reference_format_start(bib)
-        id = multi_bibitem_ref_code(bib)
-        id1 = render_multi_identifiers(id)
-        out = id1
-        date = bib.at(ns("./date[@type = 'published']/on | " \
-          "./date[@type = 'published']/from")) and
-          out << " (#{date.text.sub(/-.*$/, '')})"
-        out += ", " if date || !id1.empty?
-        out
       end
 
       def titlecase(str)
@@ -163,39 +142,6 @@ module IsoDoc
         @meta.ip_notice_received isoxml, out
         @meta.techreport isoxml, out
         super
-      end
-
-      def bibliography_bibitem_number1(bib, idx)
-        mn = bib.at(ns(".//docidentifier[@type = 'metanorma']")) and
-          /^\[?\d+\]?$/.match?(mn.text) and
-          mn["type"] = "metanorma-ordinal"
-        if (mn = bib.at(ns(".//docidentifier[@type = 'metanorma-ordinal']"))) &&
-            !bibliography_bibitem_number_skip(bib)
-          idx += 1
-          mn.children = "[#{idx}]"
-        end
-        idx
-      end
-
-      def bibliography_bibitem_number_skip(bibitem)
-        @xrefs.klass.implicit_reference(bibitem) ||
-          bibitem["hidden"] == "true" || bibitem.parent["hidden"] == "true"
-      end
-
-      def norm_ref_entry_code(_ordinal, idents, _ids, _standard, datefn, _bib)
-        ret = (idents[:metanorma] || idents[:ordinal] || idents[:sdo]).to_s
-        /^\[.+\]$/.match?(ret) or ret = "[#{ret}]"
-        ret += datefn
-        ret.empty? and return ret
-        ret.gsub("-", "&#x2011;").gsub(/ /, "&#xa0;")
-      end
-
-      def biblio_ref_entry_code(_ordinal, idents, _id, _standard, datefn, _bib)
-        ret = (idents[:metanorma] || idents[:ordinal] || idents[:sdo]).to_s
-        /^\[.+\]$/.match?(ret) or ret = "[#{ret}]"
-        ret += datefn
-        ret.empty? and return ret
-        ret.gsub("-", "&#x2011;").gsub(/ /, "&#xa0;")
       end
 
       def toc_title(docxml)
